Create Game 7: Asteroids
Asteroids is a relatively simple game by today's standards, but was still immensely popular during its time. 
In the game, the player controls a spaceship via four buttons: two buttons that rotate the spaceship clockwise or 
counterclockwise (independent of its current velocity), a thrust button that accelerates the ship in its forward 
direction and a fire button that shoots missiles. Large asteroids spawn randomly on the screen with random velocities. 
The player's goal is to destroy these asteroids before they strike the player's ship. 
In the arcade version, a large rock hit by a missile split into several fast moving small asteroids that themselves 
must be destroyed. Occasionally, a flying saucer also crosses the screen and attempts to destroy the player's spaceship. 
Searching for "asteroids arcade" yields links to multiple versions of Asteroids that are available on the web 
(including an updated version by Atari, the original creator of Asteroids).

Visit the following link to run the game...
http://www.codeskulptor.org/#user16_Wn8QGdbzgS_1.py